+++ statham/dsl/elements/base.py
@@ -5,7 +5,7 @@ from typing import Any, cast, Dict, List, Generic, TypeVar, Union
 from statham.dsl.constants import NotPassed, Maybe
 from statham.dsl.exceptions import ValidationError
 from statham.dsl.helpers import custom_repr
-from statham.dsl.property import _Property
+from statham.dsl.property import _Property, _PropertyDict
 from statham.dsl.validation import (
     get_validators,
     InstanceOf,
@@ -120,6 +120,8 @@ class Element(Generic[T]):
     """
     # pylint: enable=line-too-long
 
+    _properties: Maybe[_PropertyDict]
+
     # This is how many options there are!
     # pylint: disable=too-many-locals
     def __init__(
@@ -175,23 +177,26 @@ class Element(Generic[T]):
         self.minLength = minLength
         self.maxLength = maxLength
         self.required = required
-        self.properties = properties
-        if isinstance(self.properties, dict):
-            for name, prop in (self.properties or {}).items():
-                prop.bind(name=name, parent=self)
-                if prop.required:
-                    self.required = list(
-                        set(
-                            cast(List[str], self.required or [])
-                            + [prop.source or name]
-                        )
-                    )
+        # https://github.com/python/mypy/issues/3004
+        self.properties = properties  # type: ignore
         self.patternProperties = patternProperties
         self.additionalProperties = additionalProperties
         self.minProperties = minProperties
         self.maxProperties = maxProperties
         self.propertyNames = propertyNames
         self.dependencies = dependencies
+
+    @property
+    def properties(self) -> Maybe[_PropertyDict]:
+        return self._properties
+
+    @properties.setter
+    def properties(self, value: Maybe[Dict[str, _Property]]):
+        if isinstance(value, NotPassed):
+            self._properties = value
+            return
+        self._properties = _PropertyDict(cast(Dict[str, _Property], value))
+        self._properties.parent = self
 
     def __repr__(self):
         """Dynamically construct the repr to match value instantiation."""
@@ -207,7 +212,9 @@ class Element(Generic[T]):
         if not isinstance(other, self.__class__):
             return False
         pub_vars = lambda x: {
-            k: v for k, v in vars(x).items() if not k.startswith("_")
+            k: v
+            for k, v in vars(x).items()
+            if not k.startswith("_") or k == "_properties"
         }
         return pub_vars(self) == pub_vars(other)